backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
pydantic[email]>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
schedule>=1.2.0
numpy>=1.24.0
slowapi>=0.1.9
psycopg2-binary>=2.9.0
